move create section view to utils
